restructured notes Code into Markdown and added notes
