+++ src/MongoDB/MongoDBNotes1.tsx
@@ -1,0 +1,349 @@
+import { ListItem } from '../Components/Types';
+
+  const items: ListItem[] = [
+
+  {
+    content: [
+      { text: 'NoSQL', emphasis: true },
+      { text: 'databases like', emphasis: false },
+      { text: 'MongoDB', emphasis: true },
+      { text: 'are a', emphasis: false },
+      { text: 'non-relational', emphasis: true },
+      { text: 'database or', emphasis: false },
+      { text: 'non-relational database management system (DBMS)', emphasis: true },
+    ],
+  },
+  {
+    content: [
+      { text: 'MongoDB', emphasis: true },
+      { text: 'is', emphasis: false },
+      { text: 'open-source', emphasis: true },
+      { text: 'and uses', emphasis: false },
+      { text: 'documents,', emphasis: true },
+      { text: 'objects,', emphasis: true },
+      { text: 'or', emphasis: true },
+      { text: 'arrays,', emphasis: true },
+      { text: 'instead of', emphasis: false },
+      { text: 'tables', emphasis: true },
+      { text: 'and', emphasis: false },
+      { text: 'rows', emphasis: true },
+      { text: 'to store', emphasis: false },
+      { text: 'data', emphasis: true },
+    ],
+  },
+  {
+    content: [
+      { text: 'Documents', emphasis: true },
+      { text: 'in', emphasis: false },
+      { text: 'MongoDB', emphasis: true },
+      { text: 'are equivalent to the', emphasis: false },
+      { text: 'rows', emphasis: true },
+      { text: 'of', emphasis: false },
+      { text: 'data', emphasis: true },
+      { text: 'in', emphasis: false },
+      { text: 'SQL', emphasis: true },
+    ],
+  },
+  {
+    content: [
+      { text: 'MongoDB', emphasis: true },
+      { text: 'uses', emphasis: false },
+      { text: 'documents', emphasis: true },
+      { text: 'and', emphasis: false },
+      { text: 'objects', emphasis: true },
+      { text: 'or', emphasis: false },
+      { text: 'arrays,', emphasis: true },
+      { text: 'instead of', emphasis: false },
+      { text: 'Tables', emphasis: true },
+      { text: 'and', emphasis: false },
+      { text: 'Rows', emphasis: true },
+      { text: 'to store', emphasis: false },
+      { text: 'data', emphasis: true },
+      { text: 'which are used by', emphasis: false },
+      { text: 'SQL', emphasis: true },
+    ],
+  },
+  {
+    content: [
+      { text: 'Objects', emphasis: true },
+      { text: 'in', emphasis: false },
+      { text: 'MongoDB', emphasis: true },
+      { text: 'are stored in a format called', emphasis: false },
+      { text: 'BSON', emphasis: true },
+      { text: 'which is also known as', emphasis: false },
+      { text: 'Binary JSON', emphasis: true },
+    ],
+  },
+  {
+    content: [
+      { text: 'BSON', emphasis: true },
+      { text: 'looks and acts like', emphasis: false },
+      { text: 'JSON', emphasis: true },
+      { text: 'but is', emphasis: false },
+      { text: 'optimized', emphasis: true },
+      { text: 'for', emphasis: false },
+      { text: 'faster parsing', emphasis: true },
+    ],
+  },
+  {
+    content: [
+      { text: 'BSON', emphasis: true },
+      { text: 'is', emphasis: false },
+      { text: 'structured', emphasis: true },
+      { text: 'into', emphasis: false },
+      { text: 'key-value pairs', emphasis: true },
+    ],
+  },
+  {
+    content: [
+      { text: 'Groups of', emphasis: false },
+      { text: 'documents', emphasis: true },
+      { text: 'are stored in', emphasis: false },
+      { text: 'collections', emphasis: true },
+    ],
+  },
+  {
+    content: [
+      { text: 'Collections', emphasis: true },
+      { text: 'do', emphasis: false },
+      { text: 'NOT', emphasis: true },
+      { text: 'have an', emphasis: false },
+      { text: 'enforced schema', emphasis: true },
+      { text: 'like SQL tables to ', emphasis: false },
+      { text: 'limit', emphasis: true },
+      { text: 'the type of', emphasis: false },
+      { text: 'data', emphasis: true },
+      { text: 'that can be stored', emphasis: false },
+    ],
+  },
+  {
+    content: [
+      { text: 'Each', emphasis: false },
+      { text: 'document', emphasis: true },
+      { text: 'in a', emphasis: false },
+      { text: 'collection', emphasis: true },
+      { text: 'can have different', emphasis: false },
+      { text: 'fields (keys)', emphasis: true },
+      { text: 'and can contain', emphasis: false },
+      { text: 'different types', emphasis: true },
+      { text: 'of', emphasis: false },
+      { text: 'data', emphasis: true },
+    ],
+  },
+  {
+    content: [
+      { text: 'Related data', emphasis: true },
+      { text: 'is', emphasis: false },
+      { text: 'NOT', emphasis: true },
+      { text: 'split between tables. Instead, it is', emphasis: false },
+      { text: 'nested', emphasis: true },
+      { text: 'within', emphasis: false },
+      { text: 'objects', emphasis: true },
+      { text: 'in an', emphasis: false },
+      { text: 'embedded documents', emphasis: true },
+    ],
+  },
+  {
+    content: [
+      { text: 'You can have', emphasis: false },
+      { text: 'documents', emphasis: true },
+      { text: 'in', emphasis: false },
+      { text: 'separate collections', emphasis: true },
+      { text: 'and have something in', emphasis: false },
+      { text: 'common', emphasis: true },
+      { text: 'between them that you could use to', emphasis: false },
+      { text: 'match', emphasis: true },
+      { text: 'them together', emphasis: false },
+    ],
+  },
+  {
+    content: [
+      { text: 'Related data', emphasis: true },
+      { text: 'will be', emphasis: false },
+      { text: 'nested', emphasis: true },
+      { text: 'in an', emphasis: false },
+      { text: 'embedded document', emphasis: true },
+      { text: 'as an', emphasis: false },
+      { text: 'object', emphasis: true },
+      { text: 'or an', emphasis: false },
+      { text: 'array', emphasis: true },
+      { text: 'of', emphasis: false },
+      { text: 'objects', emphasis: true },
+
+    ],
+  },
+  {
+    content: [
+      { text: 'MongoDB', emphasis: true },
+      { text: 'can be used for', emphasis: false },
+      { text: 'heterogeneous data', emphasis: true },
+      { text: 'and is easily', emphasis: false },
+      { text: 'scalable', emphasis: true },
+    ],
+  },
+  {
+    content: [
+      { text: 'MongoDB', emphasis: true },
+      { text: 'is used to', emphasis: false },
+      { text: 'store', emphasis: true },
+      { text: 'and', emphasis: false },
+      { text: 'process', emphasis: true },
+      { text: 'data in a', emphasis: false },
+      { text: 'flexible, scalable,', emphasis: true },
+      { text: 'and', emphasis: false },
+      { text: 'distributed', emphasis: true },
+      { text: 'manner', emphasis: false },
+    ],
+  },
+  {
+    content: [
+      { text: 'Mongoose', emphasis: true },
+      { text: 'provides a', emphasis: false },
+      { text: 'light', emphasis: true },
+      { text: 'schema', emphasis: true, emphasis2: true },
+      { text: 'or', emphasis: false },
+      { text: 'lightweight', emphasis: true },
+      { text: 'wrapper', emphasis: true, emphasis2: true },
+      { text: 'on top of', emphasis: false },
+      { text: 'MongoDB', emphasis: true },
+      { text: 'and the tools that', emphasis: false },
+      { text: 'MongoDB', emphasis: true },
+      { text: 'provides', emphasis: false },
+    ],
+  },
+  {
+    content: [
+      { text: 'The relationship between', emphasis: false },
+      { text: 'Mongo', emphasis: true },
+      { text: 'and', emphasis: false },
+      { text: 'Mongoose', emphasis: true },
+      { text: 'is similar to the relationship between', emphasis: false },
+      { text: 'SQL', emphasis: true },
+      { text: 'and', emphasis: false },
+      { text: 'Sequelize', emphasis: true },
+    ],
+  },
+  {
+    content: [
+      { text: 'Mongoose', emphasis: true },
+      { text: 'enforces a more rigid', emphasis: false },
+      { text: 'structure', emphasis: true },
+      { text: 'to MongoDB', emphasis: false },
+      { text: 'similar', emphasis: true },
+      { text: 'to how', emphasis: false },
+      { text: 'TypeScript', emphasis: true },
+      { text: 'enforces', emphasis: false },
+      { text: 'types', emphasis: true },
+      { text: 'onto', emphasis: false },
+      { text: 'JavaScript', emphasis: true },
+    ],
+  },
+  {
+    content: [
+      { text: 'MongoDB Compass', emphasis: true },
+      { text: 'is the', emphasis: false },
+      { text: 'Graphic User Interface (GUI)', emphasis: true },
+      { text: 'for', emphasis: false },
+      { text: 'MongoDB', emphasis: true },
+      { text: 'which allows you to', emphasis: false },
+      { text: 'interact', emphasis: true },
+      { text: 'with your', emphasis: false },
+      { text: 'database', emphasis: true },
+      { text: '/', emphasis: false },
+      { text: 'data', emphasis: true,},
+      { text: 'in a', emphasis: false },
+      { text: 'visual', emphasis: true },
+      { text: 'manner', emphasis: false },
+    ],
+  },
+  {
+    content: [
+      { text: 'The', emphasis: false },
+      { text: 'default port', emphasis: true },
+      { text: 'for', emphasis: false },
+      { text: 'MongoDB', emphasis: true },
+      { text: 'is', emphasis: false },
+      { text: '27017', emphasis: true },
+    ],
+  },
+  {
+    content: [
+      { text: 'If you go into the', emphasis: false },
+      { text: 'Mongo shell', emphasis: true },
+      { text: 'there is a small', emphasis: false },
+      { text: `"i"`, emphasis: true },
+      { text: 'icon in the', emphasis: false },
+      { text: 'top right', emphasis: true },
+      { text: 'corner, which will pull up a', emphasis: false },
+      { text: 'list', emphasis: true },
+      { text: 'of', emphasis: false },
+      { text: 'helpful', emphasis: true },
+      { text: 'commands/keyboard shortcuts', emphasis: false },
+    ]
+  },
+  {
+    content: [
+      { text: 'The', emphasis: false },
+      { text: 'ObjectID()', emphasis: true },
+      { text: 'method', emphasis: true, emphasis2: true },
+      { text: 'is a total of', emphasis: false },
+      { text: '12 bytes', emphasis: true },
+      { text: 'long and the following:', emphasis: false},
+      { text: 'A', emphasis: false},
+      { text: '4-byte', emphasis: true },
+      { text: 'timestamp', emphasis: true, emphasis2: true },
+      { text: 'representing the', emphasis: false },
+      { text: `ObjectId's`, emphasis: true },
+      { text: 'creation', emphasis: false },
+      { text: 'A', emphasis: false },
+      { text: '5-byte', emphasis: true },
+      { text: 'random value', emphasis: true, emphasis2: true },
+      { text: 'generated', emphasis: false },
+      { text: 'once', emphasis: true },
+      { text: 'per', emphasis: false },
+      { text: 'process', emphasis: true },
+      { text: 'A', emphasis: false },
+      { text: '3-byte', emphasis: true },
+      { text: 'incrementing counter', emphasis: true, emphasis2: true },
+    ],
+  },
+  {
+    content: [
+      { text: 'The', emphasis: false },
+      { text: '4-byte', emphasis: true },
+      { text: 'timestamp ', emphasis: true, emphasis2: true },
+      { text: 'is measured in', emphasis: false },
+      { text: 'seconds', emphasis: true },
+    ],
+  },
+  {
+    content: [
+      { text: 'The', emphasis: false },
+      { text: '5-byte', emphasis: false },
+      { text: 'random value', emphasis: true, emphasis2: true },
+      { text: 'is', emphasis: false },
+      { text: 'generated', emphasis: true },
+      { text: 'once', emphasis: true, emphasis2: true },
+      { text: 'per', emphasis: false },
+      { text: 'process', emphasis: true},
+    ],
+  },
+  {
+    content: [
+      { text: '', emphasis: true },
+      { text: '', emphasis: false },
+      { text: '', emphasis: true },
+      { text: '', emphasis: false },
+    ],
+  },
+  {
+    content: [
+      { text: '', emphasis: true },
+      { text: '', emphasis: false },
+      { text: '', emphasis: true },
+      { text: '', emphasis: false },
+    ],
+  },
+];
+
+export default items;
