+++ src/ReactNotes/ReactNotes.tsx
@@ -151,14 +151,6 @@ const ReactItems: ListItem[] = [
       { text: '', emphasis: false },
     ],
   },
-  {
-    content: [
-      { text: '', emphasis: true },
-      { text: '', emphasis: false },
-      { text: '', emphasis: true },
-      { text: '', emphasis: false },
-    ],
-  },
 ];
 
 export default ReactItems;
