+++ src/pages/API.tsx
@@ -1,5 +1,5 @@
 import Navigation from "../Components/Navigation";
-import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
+// import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
 import apiItemsWithCode from "../APINotes/APINotesWCode";
 import "../CSS/Header.css";
 import APINotes from "../APINotes/APINotes";
@@ -10,8 +10,8 @@ export default function APIs() {
         <>
             <h1 className="Header">APIs</h1>
             <Navigation />
-            <DynamicNotes items={APINotes} />
-            <DynamicNotesCode items={apiItemsWithCode} />
+            {/* <DynamicNotes items={APINotes} />
+            <DynamicNotesCode items={apiItemsWithCode} /> */}
         </>
     );
