+++ src/pages/CSS.tsx
@@ -1,5 +1,5 @@
 import Navigation from "../Components/Navigation";
-import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
+// import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
 import items from "../CSSNotes/CSSNotes";
 import cssItemsWithCode from "../CSSNotes/CSSCodeExamples";
 import "../CSS/Header.css";
@@ -10,8 +10,8 @@ export default function css() {
         <>
             <h1 className="Header">CSS</h1>
             <Navigation />
-            <DynamicNotes items={items} />
-            <DynamicNotesCode items={cssItemsWithCode} />
+            {/* <DynamicNotes items={items} />
+            <DynamicNotesCode items={cssItemsWithCode} /> */}
 
         </>
     );
