+++ src/pages/FullStackReact.tsx
@@ -1,9 +1,9 @@
 // import DynamicNotes, { DynamicNotesCode } from "../Notes/DynamicNotes";
 import Navigation from "../Components/Navigation";
 import "../CSS/Header.css";
-import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
-import items from "../ReactFullStackNotes/FullStackReactNotes";
-import fullStackReactItemsWithCode from "../ReactFullStackNotes/FullStackReactNotesWCode";
+// import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
+// import items from "../ReactFullStackNotes/FullStackReactNotes";
+// import fullStackReactItemsWithCode from "../ReactFullStackNotes/FullStackReactNotesWCode";
 export default function fullStackReact() {
 
     return (
@@ -11,8 +11,8 @@ export default function fullStackReact() {
         <>
             <h1 className="Header">Full-Stack React</h1>
             <Navigation />
-            <DynamicNotes items={items} />
-            <DynamicNotesCode items={fullStackReactItemsWithCode} />
+            {/* <DynamicNotes items={items} />
+            <DynamicNotesCode items={fullStackReactItemsWithCode} /> */}
 
         </>
     );
