+++ src/pages/Git.tsx
@@ -1,8 +1,8 @@
-import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
+// import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
 import Navigation from "../Components/Navigation";
 import "../CSS/Header.css";
-import GitItems from "../GitNotes/GitNotes";
-import GitItemsWithCode from "../GitNotes/GitNotesWithCode";
+// import GitItems from "../GitNotes/GitNotes";
+// import GitItemsWithCode from "../GitNotes/GitNotesWithCode";
 export default function git() {
 
     return (
@@ -10,8 +10,8 @@ export default function git() {
         <>
             <h1 className="Header">Git</h1>
             <Navigation />
-            <DynamicNotes items={ GitItems } />
-            <DynamicNotesCode items={ GitItemsWithCode } />
+            {/* <DynamicNotes items={ GitItems } />
+            <DynamicNotesCode items={ GitItemsWithCode } /> */}
 
         </>
     );
