+++ src/pages/HTML.tsx
@@ -1,8 +1,8 @@
-import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
-import htmlItemsWithCode from "../HTMLNotes/HTMLCodeExamples";
+// import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
+// import htmlItemsWithCode from "../HTMLNotes/HTMLCodeExamples";
 import Navigation from "../Components/Navigation";
 import "../CSS/Header.css";
-import HTMLItems from "../HTMLNotes/HTMLNotes";
+// import HTMLItems from "../HTMLNotes/HTMLNotes";
 
 
 export default function html() {
@@ -12,8 +12,8 @@ export default function html() {
         <>
             <h1 className="Header">HTML</h1>
             <Navigation />
-            <DynamicNotes items={ HTMLItems } />
-            <DynamicNotesCode items={htmlItemsWithCode} />
+            {/* <DynamicNotes items={ HTMLItems } />
+            <DynamicNotesCode items={htmlItemsWithCode} /> */}
         </>
     );
