+++ src/pages/JavaScript.tsx
@@ -1,7 +1,7 @@
-import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
+// import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
 import Navigation from "../Components/Navigation";
-import JavaScriptItems from "../JavaScriptNotes/JavaScriptNotes";
-import javascriptItemsWithCode from "../JavaScriptNotes/JavaScriptNotesWithCode";
+// import JavaScriptItems from "../JavaScriptNotes/JavaScriptNotes";
+// import javascriptItemsWithCode from "../JavaScriptNotes/JavaScriptNotesWithCode";
 import "../CSS/Header.css";
 export default function javascript() {
 
@@ -10,8 +10,8 @@ export default function javascript() {
         <>
             <h1 className="Header">JavaScript</h1>
             <Navigation />
-            <DynamicNotes items={ JavaScriptItems } />
-            <DynamicNotesCode items={ javascriptItemsWithCode } />
+            {/* <DynamicNotes items={ JavaScriptItems } />
+            <DynamicNotesCode items={ javascriptItemsWithCode } /> */}
         </>
     );
