+++ src/pages/MernAndAuth.tsx
@@ -1,8 +1,8 @@
 import Navigation from "../Components/Navigation";
-import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
+// import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
 import "../CSS/Header.css";
-import MernAndAuthItems from "../MernAndAuth/MernAndAuthNotes.tsx";
-import MERNItemsWithCode from "../MernAndAuth/MernAndAuthNotesWCode.tsx";
+// import MernAndAuthItems from "../MernAndAuth/MernAndAuthNotes.tsx";
+// import MERNItemsWithCode from "../MernAndAuth/MernAndAuthNotesWCode.tsx";
 
 
 export default function MernAndAuth() {
@@ -12,8 +12,8 @@ export default function MernAndAuth() {
         <>
             <h1 className="Header">MERN Stack</h1>
             <Navigation />
-            <DynamicNotes items={ MernAndAuthItems } />
-            <DynamicNotesCode items={ MERNItemsWithCode } />
+            {/* <DynamicNotes items={ MernAndAuthItems } />
+            <DynamicNotesCode items={ MERNItemsWithCode } /> */}
         </>
     );
 }
