+++ src/pages/Node.tsx
@@ -1,7 +1,7 @@
-import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
+// import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
 import Navigation from "../Components/Navigation";
-import items from "../NodeJSNotes/NodeJSNotes";
-import nodeItemsWithCode from "../NodeJSNotes/NodeJSNotesWithCode";
+// import items from "../NodeJSNotes/NodeJSNotes";
+// import nodeItemsWithCode from "../NodeJSNotes/NodeJSNotesWithCode";
 import "../CSS/Header.css";
 export default function nodeJS() {
 
@@ -10,8 +10,8 @@ export default function nodeJS() {
         <>
             <h1 className="Header">Node JS</h1>
             <Navigation />
-            <DynamicNotes items={items} />
-            <DynamicNotesCode items={nodeItemsWithCode} />
+            {/* <DynamicNotes items={items} />
+            <DynamicNotesCode items={nodeItemsWithCode} /> */}
         </>
     );
