+++ src/pages/PostgreSQL.tsx
@@ -1,7 +1,7 @@
-import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
+// import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
 import Navigation from "../Components/Navigation";
-import items from "../PostgreSQL/PostgreSQLNotes.tsx";
-import postgreSQLItemsWithCode from "../PostgreSQL/PostgreSQLNotesWCode.tsx";
+// import items from "../PostgreSQL/PostgreSQLNotes.tsx";
+// import postgreSQLItemsWithCode from "../PostgreSQL/PostgreSQLNotesWCode.tsx";
 import "../CSS/Header.css";
 
 
@@ -12,8 +12,8 @@ export default function postgreSQL() {
         <>
             <h1 className="Header">PostgreSQL</h1>
             <Navigation />
-            <DynamicNotes items={items} />
-            <DynamicNotesCode items={postgreSQLItemsWithCode} />
+            {/* <DynamicNotes items={items} />
+            <DynamicNotesCode items={postgreSQLItemsWithCode} /> */}
         </>
     );
