+++ src/pages/Python.tsx
@@ -1,8 +1,8 @@
 import Navigation from "../Components/Navigation";
 import "../CSS/Header.css";
-import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
-import pythonItemsWithCode from "../PythonNotes/PythonNotesWithCode";
-import items from "../PythonNotes/PythonNotes";
+// import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
+// import pythonItemsWithCode from "../PythonNotes/PythonNotesWithCode";
+// import items from "../PythonNotes/PythonNotes";
 
 
 export default function Python() {
@@ -12,8 +12,8 @@ export default function Python() {
         <>
             <h1 className="Header">MongoDB</h1>
             <Navigation />
-            <DynamicNotes items={ items } />
-            <DynamicNotesCode items={ pythonItemsWithCode } />
+            {/* <DynamicNotes items={ items } />
+            <DynamicNotesCode items={ pythonItemsWithCode } /> */}
 
         </>
     );
