+++ src/pages/React.tsx
@@ -1,8 +1,8 @@
 import Navigation from "../Components/Navigation";
-import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
+// import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
 import "../CSS/Header.css";
-import ReactItems from "../ReactNotes/ReactNotes";
-import ReactItemsWithCode from "../ReactNotes/ReactCodeExamples";
+// import ReactItems from "../ReactNotes/ReactNotes";
+// import ReactItemsWithCode from "../ReactNotes/ReactCodeExamples";
 
 
 export default function react() {
@@ -12,8 +12,8 @@ export default function react() {
         <>
             <h1 className="Header">React</h1>
             <Navigation />
-            <DynamicNotes items={ ReactItems } />
-            <DynamicNotesCode items={ ReactItemsWithCode } />
+            {/* <DynamicNotes items={ ReactItems } />
+            <DynamicNotesCode items={ ReactItemsWithCode } /> */}
         </>
     );
