+++ src/pages/ReactandTypeScript.tsx
@@ -1,7 +1,7 @@
-import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
+// import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
 import Navigation from "../Components/Navigation";
-import items from "../ReactWithTS/ReactWTypeScriptNotes";
-import reactAndTypescriptItemsWithCode from "../ReactWithTS/ReactWTypeScriptNotesWithCode";
+// import items from "../ReactWithTS/ReactWTypeScriptNotes";
+// import reactAndTypescriptItemsWithCode from "../ReactWithTS/ReactWTypeScriptNotesWithCode";
 import "../CSS/Header.css";
 export default function reactAndTypeScript() {
 
@@ -10,8 +10,8 @@ export default function reactAndTypeScript() {
         <>
             <h1 className="Header">React + TypeScript</h1>
             <Navigation />
-            <DynamicNotes items={items} />
-            <DynamicNotesCode items={reactAndTypescriptItemsWithCode} />
+            {/* <DynamicNotes items={items} />
+            <DynamicNotesCode items={reactAndTypescriptItemsWithCode} /> */}
         </>
     );
 }
