+++ src/pages/Servers.tsx
@@ -1,8 +1,8 @@
-import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
+// import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
 import Navigation from "../Components/Navigation";
 import "../CSS/Header.css";
-import serversNotes from "../Servers/ServersNotes";
-import serversNotesWCode from "../Servers/SeversNotesWCode";
+// import serversNotes from "../Servers/ServersNotes";
+// import serversNotesWCode from "../Servers/SeversNotesWCode";
 export default function servers() {
 
     return (
@@ -10,8 +10,8 @@ export default function servers() {
         <>
             <h1 className="Header">Servers</h1>
             <Navigation />
-            <DynamicNotes items={serversNotes} />
-            <DynamicNotesCode items={serversNotesWCode} />
+            {/* <DynamicNotes items={serversNotes} />
+            <DynamicNotesCode items={serversNotesWCode} /> */}
         </>
     );
 }
