+++ src/pages/TypeScript.tsx
@@ -1,8 +1,8 @@
 import Navigation from "../Components/Navigation";
-import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
-import typescriptItemsWithCode from "../TypeScriptNotes/TypeScriptNotesWithCode";
+// import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
+// import typescriptItemsWithCode from "../TypeScriptNotes/TypeScriptNotesWithCode";
 import "../CSS/Header.css";
-import TypeScriptItems from "../TypeScriptNotes/TypeScriptNotes";
+// import TypeScriptItems from "../TypeScriptNotes/TypeScriptNotes";
 
 export default function typescript() {
 
@@ -11,8 +11,8 @@ export default function typescript() {
         <>
             <h1 className="Header">TypeScript</h1>
             <Navigation />
-            <DynamicNotes items={ TypeScriptItems } />
-            <DynamicNotesCode items={typescriptItemsWithCode} />
+            {/* <DynamicNotes items={ TypeScriptItems } />
+            <DynamicNotesCode items={typescriptItemsWithCode} /> */}
         </>
     );
