+++ src/pages/WebAPI.tsx
@@ -1,7 +1,7 @@
-import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
+// import DynamicNotes, { DynamicNotesCode } from "../Components/DynamicNotes";
 import Navigation from "../Components/Navigation";
-import items from "../Notes/WebAPINotes";
-import itemsWithCode from "../Notes/WebAPINotesWithCode";
+// import items from "../Notes/WebAPINotes";
+// import itemsWithCode from "../Notes/WebAPINotesWithCode";
 import "../CSS/Header.css";
 export default function webAPI() {
 
@@ -10,8 +10,8 @@ export default function webAPI() {
         <>
             <h1 className="Header">APIs</h1>
             <Navigation />
-            <DynamicNotes items={items} />
-            <DynamicNotesCode items={itemsWithCode} />
+            {/* <DynamicNotes items={items} />
+            <DynamicNotesCode items={itemsWithCode} /> */}
         </>
     );
 }
