--- stackbit.config.ts
@@ -1,9 +1,0 @@
-import { defineStackbitConfig } from '@stackbit/types';
-
-export default defineStackbitConfig({
-    "stackbitVersion": "~0.6.0",
-    "nodeVersion": "18",
-    "ssgName": "custom",
-    "contentSources": [],
-    "postInstallCommand": "npm i --no-save @stackbit/types"
-})
